+++ packages/clients/snowbox/src/addPlugins.ts
@@ -11,7 +11,7 @@ import IconMenu from '@polar/plugin-icon-menu'
 import LayerChooser from '@polar/plugin-layer-chooser'
 import Legend from '@polar/plugin-legend'
 import LoadingIndicator from '@polar/plugin-loading-indicator'
-import MousePosition from '@polar/plugin-mouse-position'
+import PointerPosition from '@polar/plugin-pointer-position'
 import Pins from '@polar/plugin-pins'
 import ReverseGeocoder from '@polar/plugin-reverse-geocoder'
 import Scale from '@polar/plugin-scale'
@@ -119,7 +119,7 @@ export const addPlugins = (core) => {
       addressTarget: 'plugin/addressSearch/selectResult',
       zoomTo: 7,
     }),
-    MousePosition(
+    PointerPosition(
       merge({}, defaultOptions, {
         layoutTag: NineLayoutTag.BOTTOM_LEFT,
       })