+++ packages/plugins/MousePosition/src/components/index.ts
@@ -1,1 +1,0 @@
-export { default as MousePosition } from './MousePosition.vue'